Extract Labeling and Selector context
Instead of cramming everything in the Devices context, extract the Selector to
its own top level context. Contextually, also add a Labeling context that will
handle tags and attributes.
