+++ backend/lib/edgehog/devices/device.ex
@@ -23,7 +23,8 @@ defmodule Edgehog.Devices.Device do
   import Ecto.Changeset
 
   alias Edgehog.Astarte
-  alias Edgehog.Devices.{Attribute, DeviceTag, SystemModelPartNumber, Tag}
+  alias Edgehog.Devices.SystemModelPartNumber
+  alias Edgehog.Labeling
 
   schema "devices" do
     field :tenant_id, :integer, autogenerate: {Edgehog.Repo, :get_tenant_id, []}
@@ -41,9 +42,9 @@ defmodule Edgehog.Devices.Device do
       type: :string
 
     has_one :system_model, through: [:system_model_part_number, :system_model]
-    many_to_many :tags, Tag, join_through: DeviceTag, on_replace: :delete
+    many_to_many :tags, Labeling.Tag, join_through: Labeling.DeviceTag, on_replace: :delete
 
-    has_many :custom_attributes, Attribute,
+    has_many :custom_attributes, Labeling.DeviceAttribute,
       where: [namespace: "custom"],
       on_replace: :delete
 
@@ -55,6 +56,6 @@ defmodule Edgehog.Devices.Device do
     device
     |> cast(attrs, [:name])
     |> validate_required([:name])
-    |> cast_assoc(:custom_attributes, with: &Attribute.custom_attribute_changeset/2)
+    |> cast_assoc(:custom_attributes, with: &Labeling.DeviceAttribute.custom_attribute_changeset/2)
   end
 end
