+++ backend/lib/edgehog/labeling/tag.ex
@@ -18,24 +18,22 @@
 # SPDX-License-Identifier: Apache-2.0
 #
 
-defmodule Edgehog.Devices.DeviceTag do
+defmodule Edgehog.Labeling.Tag do
   use Ecto.Schema
-  import Ecto.Query
-  alias Edgehog.Devices.DeviceTag
-  alias Edgehog.Devices.Tag
+  import Ecto.Changeset
 
-  @primary_key false
-  schema "devices_tags" do
+  schema "tags" do
     field :tenant_id, :integer, autogenerate: {Edgehog.Repo, :get_tenant_id, []}
-    field :tag_id, :id, primary_key: true
-    field :device_id, :id, primary_key: true
+    field :name, :string
+
+    timestamps()
   end
 
-  def device_ids_matching_tag(tag) when is_binary(tag) do
-    from dt in DeviceTag,
-      join: t in Tag,
-      on: dt.tag_id == t.id,
-      where: t.name == ^tag,
-      select: dt.device_id
+  @doc false
+  def changeset(tag, attrs) do
+    tag
+    |> cast(attrs, [:name])
+    |> validate_required([:name])
+    |> unique_constraint([:name, :tenant_id])
   end
 end
