+++ backend/lib/edgehog/selector/ast/binary_op.ex
@@ -18,6 +18,6 @@
 # SPDX-License-Identifier: Apache-2.0
 #
 
-defmodule Edgehog.Devices.Selector.AST.BinaryOp do
+defmodule Edgehog.Selector.AST.BinaryOp do
   defstruct [:operator, :lhs, :rhs]
 end
